Enable StatsD performance monitoring on .NET Standard
This is probably a binary breaking change (due to changing class definitions), however:

StatsD integration will work when on .NET Core / .NET Standard environments
